+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   undefined custom cache) when parsing step caches (`--show`, running a
   pipeline step etc.).
 - php cs-fixer custom fixers (thanks [Kuba Werlos])
+### Changed
+- improve [pipeline cache docs](doc/PIPELINES-CACHES.md)
 
 [Kuba Werlos]: https://github.com/kubawerlos